requirements.txt
#
# This file is autogenerated by pip-compile with python 3.10
# To update, run:
#
#    pip-compile
#
aiohttp==3.8.1
    # via web3
aiosignal==1.2.0
    # via aiohttp
ape-vyper==0.3.0
    # via -r requirements.in
appnope==0.1.3
    # via ipython
asttokens==2.0.5
    # via stack-data
async-timeout==4.0.2
    # via aiohttp
attrs==21.4.0
    # via
    #   aiohttp
    #   jsonschema
    #   pytest
backcall==0.2.0
    # via ipython
base58==1.0.3
    # via
    #   multiaddr
    #   py-cid
    #   py-multihash
bitarray==1.2.2
    # via eth-account
cached-property==1.5.2
    # via
    #   py-ecc
    #   py-evm
certifi==2022.6.15
    # via requests
cffi==1.15.1
    # via
    #   pygit2
    #   pynacl
charset-normalizer==2.1.0
    # via
    #   aiohttp
    #   requests
click==8.1.3
    # via eth-ape
colorama==0.4.5
    # via rich
commonmark==0.9.1
    # via rich
cytoolz==0.12.0
    # via
    #   eth-keyfile
    #   eth-utils
decorator==5.1.1
    # via ipython
deprecated==1.2.13
    # via pygithub
eth-abi==2.2.0
    # via
    #   eth-account
    #   eth-ape
    #   eth-tester
    #   web3
eth-account==0.5.7
    # via
    #   eth-ape
    #   eth-tester
    #   web3
eth-ape==0.3.5
    # via
    #   -r requirements.in
    #   ape-vyper
eth-bloom==1.0.4
    # via py-evm
eth-hash[pycryptodome,pysha3]==0.3.3
    # via
    #   eth-bloom
    #   eth-tester
    #   eth-utils
    #   trie
    #   web3
eth-keyfile==0.5.1
    # via eth-account
eth-keys==0.3.4
    # via
    #   eth-account
    #   eth-keyfile
    #   eth-tester
    #   py-evm
eth-rlp==0.2.1
    # via
    #   eth-account
    #   eth-ape
    #   web3
eth-tester[py-evm]==v0.6.0-beta.6
    # via web3
eth-typing==2.3.0
    # via
    #   eth-abi
    #   eth-keys
    #   eth-utils
    #   py-ecc
    #   py-evm
    #   web3
eth-utils==1.10.0
    # via
    #   eth-abi
    #   eth-account
    #   eth-ape
    #   eth-keyfile
    #   eth-keys
    #   eth-rlp
    #   eth-tester
    #   ethpm-types
    #   evm-trace
    #   py-ecc
    #   py-evm
    #   rlp
    #   trie
    #   web3
ethpm-types==0.3.2
    # via eth-ape
evm-trace==0.1.0a6
    # via eth-ape
executing==0.8.3
    # via stack-data
frozenlist==1.3.0
    # via
    #   aiohttp
    #   aiosignal
hexbytes==0.2.2
    # via
    #   eth-account
    #   eth-ape
    #   eth-rlp
    #   ethpm-types
    #   evm-trace
    #   trie
    #   web3
idna==3.3
    # via
    #   requests
    #   yarl
importlib-metadata==4.12.0
    # via eth-ape
iniconfig==1.1.1
    # via pytest
ipfshttpclient==0.8.0a2
    # via web3
ipython==8.4.0
    # via eth-ape
jedi==0.18.1
    # via ipython
jsonschema==4.7.2
    # via web3
lru-dict==1.1.8
    # via
    #   py-evm
    #   web3
matplotlib-inline==0.1.3
    # via ipython
morphys==1.0
    # via
    #   py-cid
    #   py-multibase
    #   py-multicodec
    #   py-multihash
multiaddr==0.0.9
    # via ipfshttpclient
multidict==6.0.2
    # via
    #   aiohttp
    #   yarl
mypy-extensions==0.4.3
    # via
    #   py-ecc
    #   py-evm
netaddr==0.8.0
    # via multiaddr
numpy==1.23.1
    # via pandas
packaging==20.9
    # via
    #   eth-ape
    #   pytest
pandas==1.4.3
    # via eth-ape
parsimonious==0.8.1
    # via eth-abi
parso==0.8.3
    # via jedi
pexpect==4.8.0
    # via ipython
pickleshare==0.7.5
    # via ipython
pluggy==0.13.1
    # via
    #   eth-ape
    #   pytest
prompt-toolkit==3.0.30
    # via ipython
protobuf==3.20.1
    # via web3
ptyprocess==0.7.0
    # via pexpect
pure-eval==0.2.2
    # via stack-data
py==1.11.0
    # via pytest
py-cid==0.3.0
    # via ethpm-types
py-ecc==5.2.0
    # via py-evm
py-evm==0.5.0a3
    # via eth-tester
py-geth==3.8.0
    # via
    #   eth-ape
    #   web3
py-multibase==1.0.3
    # via py-cid
py-multicodec==0.2.1
    # via py-cid
py-multihash==0.2.3
    # via py-cid
pycparser==2.21
    # via cffi
pycryptodome==3.15.0
    # via
    #   eth-hash
    #   eth-keyfile
pydantic==1.9.1
    # via
    #   eth-ape
    #   ethpm-types
    #   evm-trace
pyethash==0.1.27
    # via py-evm
pygit2==1.9.2
    # via eth-ape
pygithub==1.55
    # via eth-ape
pygments==2.12.0
    # via
    #   ipython
    #   rich
pyjwt==2.4.0
    # via pygithub
pynacl==1.5.0
    # via pygithub
pyparsing==3.0.9
    # via packaging
pyrsistent==0.18.1
    # via jsonschema
pysha3==1.0.2
    # via eth-hash
pytest==7.1.2
    # via eth-ape
python-baseconv==1.2.2
    # via py-multibase
python-dateutil==2.8.2
    # via
    #   eth-ape
    #   pandas
pytz==2022.1
    # via pandas
pyyaml==6.0
    # via eth-ape
requests==2.28.1
    # via
    #   eth-ape
    #   ipfshttpclient
    #   pygithub
    #   vvm
    #   web3
rich==10.16.2
    # via eth-ape
rlp==2.0.1
    # via
    #   eth-account
    #   eth-rlp
    #   eth-tester
    #   py-evm
    #   trie
semantic-version==2.10.0
    # via
    #   eth-tester
    #   py-geth
    #   vvm
six==1.16.0
    # via
    #   asttokens
    #   multiaddr
    #   parsimonious
    #   py-multibase
    #   py-multicodec
    #   py-multihash
    #   python-dateutil
sortedcontainers==2.4.0
    # via trie
stack-data==0.3.0
    # via ipython
tomli==2.0.1
    # via pytest
toolz==0.12.0
    # via cytoolz
tqdm==4.64.0
    # via
    #   ape-vyper
    #   eth-ape
traitlets==5.3.0
    # via
    #   ipython
    #   matplotlib-inline
trie==2.0.0-alpha.5
    # via py-evm
typing-extensions==3.10.0.2
    # via
    #   pydantic
    #   trie
urllib3==1.26.10
    # via requests
varint==1.0.2
    # via
    #   multiaddr
    #   py-multicodec
    #   py-multihash
vvm==0.1.0
    # via ape-vyper
wcwidth==0.2.5
    # via prompt-toolkit
web3[tester]==5.30.0
    # via eth-ape
websockets==9.1
    # via web3
wrapt==1.14.1
    # via deprecated
yarl==1.7.2
    # via aiohttp
zipp==3.8.1
    # via importlib-metadata

# The following packages are considered to be unsafe in a requirements file:
# setuptools
